ertetretert

+++ www/js/com/App.js
@@ -13,9 +13,9 @@ function App(){
 	};
 	this.redirigiendo_una_push = false;
 	this.cargo_mapa = false;
-	//this.server = 'http://192.168.0.2/s4nt4nd3rs_4pp/server/'
+	this.server = 'http://192.168.0.2/s4nt4nd3rs_4pp/server/'
 	//this.server = 'http://192.168.235.140:8888/s4nt4nd3rs_4pp/server/';
-	this.server = 'http://santander.crudo.com.uy/';
+	//this.server = 'http://santander.crudo.com.uy/';
 	
 	this.db = openDatabase('santanders_app_punta', '1.0', 'santanders_app_punta', 2000000);
 	this._ManagePush;
@@ -323,16 +323,20 @@ function App(){
 				where += ' eventos_id=' + array_del[del_id] ;
 
 			}
-			
-			tx.executeSql('DELETE FROM "eventos" WHERE ' + where);
 
+			if($($xml).find('del').text() !=''){
+						
+						tx.executeSql('DELETE FROM "eventos" WHERE ' + where);
+			}
 			tx.executeSql('UPDATE app SET sync_value=?', [new_sync_value]);
 			$(document).trigger('LISTAR_EVENTOS');
+
+
 
 		},  app.db_errorGeneral);
 
 		
-
+		
 	}
